+++ src/Core/IDataLoader.cs
@@ -20,15 +20,34 @@ namespace GreenDonut
     public interface IDataLoader
     {
         /// <summary>
+        /// An event that will be raised when a incoming data request is added
+        /// to the buffer. Will never be raised if batching is disabled.
+        /// </summary>
+        event RequestBufferedEventHandler RequestBuffered;
+
+        /// <summary>
+        /// Gets the current count of buffered data requests waiting for being
+        /// dispatched as batches. Will always return <c>0</c> if batching is
+        /// disabled.
+        /// </summary>
+        int BufferedRequests { get; }
+
+        /// <summary>
+        /// Gets the current count of cached values. Will always return
+        /// <c>0</c> if caching is disabled.
+        /// </summary>
+        int CachedValues { get; }
+
+        /// <summary>
         /// Empties the complete cache.
         /// </summary>
         void Clear();
 
         /// <summary>
-        /// Dispatches one or more batch requests.
-        /// In case of auto dispatching we just trigger an implicit dispatch
-        /// which could mean to interrupt a wait delay. Whereas in a manual
-        /// dispatch scenario it could mean to dispatch explicitly.
+        /// Dispatches one or more batch requests. In case of auto dispatching
+        /// we just trigger an implicit dispatch which could mean to interrupt
+        /// a wait delay. Whereas in a manual dispatch scenario it could mean
+        /// to dispatch explicitly.
         /// </summary>
         Task DispatchAsync();